web/requirements.txt
django>=3.2,<4.0
gunicorn
django-dotenv
psycopg2-binary
djangorestframework
django-filter
markdown
django-guardian
 